TASK: Split error messages off
The error messages and especially the result object are usable on their
own and splitting them off also makes splitting the validator package off
much easier and allows it to be reused without too much dependencies.
