+++ Neos.Error.Messages/Classes/TYPO3/Flow/Error/Notice.php
@@ -2,7 +2,7 @@
 namespace TYPO3\Flow\Error;
 
 /*
- * This file is part of the TYPO3.Flow package.
+ * This file is part of the Neos.Error.Messages package.
  *
  * (c) Contributors of the Neos Project - www.neos.io
  *
@@ -12,16 +12,16 @@ namespace TYPO3\Flow\Error;
  */
 
 /**
- * An object representation of a generic error. Subclass this to create
- * more specific errors if necessary.
+ * An object representation of a generic notice. Subclass this to create
+ * more specific notices if necessary.
  *
  * @api
  */
-class Error extends Message
+class Notice extends Message
 {
     /**
-     * The severity of this message ('Error').
+     * The severity of this message ('Notice').
      * @var string
      */
-    protected $severity = self::SEVERITY_ERROR;
+    protected $severity = self::SEVERITY_NOTICE;
 }
