+++ Neos.Error.Messages/Tests/Unit/ResultTest.php
@@ -1,8 +1,8 @@
 <?php
-namespace TYPO3\Flow\Tests\Unit\Error;
+namespace Neos\Error\Messages\Tests\Unit;
 
 /*
- * This file is part of the TYPO3.Flow package.
+ * This file is part of the Neos.Error.Messages package.
  *
  * (c) Contributors of the Neos Project - www.neos.io
  *
@@ -15,7 +15,7 @@ namespace TYPO3\Flow\Tests\Unit\Error;
  * Testcase for the Error Container object
  *
  */
-class ResultTest extends \TYPO3\Flow\Tests\UnitTestCase
+class ResultTest extends \PHPUnit_Framework_TestCase
 {
     /**
      *
