+++ Neos.Error.Messages/composer.json
@@ -1,0 +1,21 @@
+{
+    "name": "neos/error-messages",
+    "type": "library",
+    "description": "Flow error messages",
+    "homepage": "https://www.flowframework.io",
+    "license": "MIT",
+
+    "require": {
+        "php": ">=5.5.0"
+    },
+    "autoload": {
+        "psr-0": {
+            "TYPO3\\Flow\\Error": "Classes"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Neos\\Error\\Messages\\Tests\\": "Tests"
+        }
+    }
+}
